+++ crates/affected/tests/affected_tracker_test.rs
@@ -1249,5 +1249,41 @@ mod affected_tasks {
                 )])
             );
         }
+
+        #[tokio::test]
+        async fn when_ci_only_tracks_when_ci() {
+            let workspace_graph = build_graph("tasks").await;
+            let touched_files = FxHashSet::from_iter(["ci/file.txt".into()]);
+
+            let mut tracker = AffectedTracker::new(workspace_graph.into(), touched_files);
+            tracker.set_ci_check(true);
+            tracker
+                .track_tasks_by_target(&[Target::parse("ci:only").unwrap()])
+                .unwrap();
+            let affected = tracker.build();
+
+            assert_eq!(
+                affected.tasks,
+                FxHashMap::from_iter([(
+                    Target::parse("ci:only").unwrap(),
+                    create_state_from_file("ci/file.txt")
+                )])
+            );
+        }
+
+        #[tokio::test]
+        async fn when_ci_only_doesnt_track_when_not_in_ci() {
+            let workspace_graph = build_graph("tasks").await;
+            let touched_files = FxHashSet::from_iter(["ci/file.txt".into()]);
+
+            let mut tracker = AffectedTracker::new(workspace_graph.into(), touched_files);
+            tracker.set_ci_check(false);
+            tracker
+                .track_tasks_by_target(&[Target::parse("ci:only").unwrap()])
+                .unwrap();
+            let affected = tracker.build();
+
+            assert!(affected.tasks.is_empty());
+        }
     }
 }